Fixed the snackbar to show the 1/2 count for stars, if it is selected in the ratingBar.

+++ MichaelModuleMid/src/main/java/michael/quon/n01565129/QuonFragment.java
@@ -57,7 +57,7 @@ public class QuonFragment extends Fragment {
         button.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                int stars = (int) ratingBar.getRating();
+                float stars = ratingBar.getRating();
                 String message = getString(R.string.rating) + stars;
                 Snackbar.make(view, message, Snackbar.LENGTH_SHORT).show();
             }
